# server/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: FMD
  description: A website for movies' and TV shows' reviews
  termsOfService: http://swagger.io/terms/
  contact:
    email: om.dehlantube@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.0.0
servers:
- url: http://localhost:8080/VEDANTANEOGI_1/FMD_API3.0/3.0.0
...
paths:
  /search:
    get:
      tags:
      - Titles
      description: Obtain all movies/shows whose prefix matches the input expression
      operationId: search_get
      parameters:
      - name: expression
        in: query
        description: the expression to be searched
        required: true
        style: form
        explode: false
        schema:
          type: string
          example: the dark knight
      responses:
        "200":
          description: Successful search
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/title'
                x-content-type: application/json
        "404":
          description: Movie/TV Show Not found
      x-openapi-router-controller: swagger_server.controllers.titles_controller
...
  /movie:
    get:
      tags:
      - Titles
      summary: Sort/Filter movies
      operationId: movie_get
      parameters:
      - name: genre
        in: query
        description: Genres to be filtered by
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      - name: sort_type_browse
        in: query
        description: Parameter based on which titles will be sorted
        required: false
        style: form
        explode: false
        schema:
          type: string
          enum:
          - Lex
          - Pop
          - Rel
          - Rat
      - name: sort_order
        in: query
        description: sorting order
        required: false
        style: form
        explode: false
        schema:
          type: boolean
          enum:
          - true
          - false
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/movie'
                x-content-type: application/json
        "400":
          description: Invalid request
        "404":
          description: No movies found
      x-openapi-router-controller: swagger_server.controllers.titles_controller
  /tvshow:
    get:
      tags:
      - Titles
      summary: Sort/Filter TV Shows
      operationId: tvshow_get
      parameters:
      - name: genre
        in: query
        description: Genres to be filtered by
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      - name: sort_type_browse
        in: query
        description: Parameter based on which titles will be sorted
        required: false
        style: form
        explode: false
        schema:
          type: string
          enum:
          - Lex
          - Pop
          - Rel
          - Rat
      - name: sort_order
        in: query
        description: sorting order
        required: false
        style: form
        explode: false
        schema:
          type: boolean
          enum:
          - true
          - false
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/tvshow'
                x-content-type: application/json
        "400":
          description: Invalid request
        "404":
          description: No TV Shows found
      x-openapi-router-controller: swagger_server.controllers.titles_controller
  /title:
    get:
      tags:
      - Titles
      summary: Sort/Filter Movies and TV Shows
      operationId: title_get
      parameters:
      - name: genre
        in: query
        description: Genres to be filtered by
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      - name: sort_type_browse
        in: query
        description: Parameter based on which titles will be sorted
        required: false
        style: form
        explode: false
        schema:
          type: string
          enum:
          - Lex
          - Pop
          - Rel
          - Rat
      - name: sort_order
        in: query
        description: sorting order
        required: false
        style: form
        explode: false
        schema:
          type: boolean
          enum:
          - true
          - false
      responses:
        "200":
          description: List of Movies/TV Shows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/title'
                x-content-type: application/json
        "400":
          description: Invalid request
        "404":
          description: No titles found
      x-openapi-router-controller: swagger_server.controllers.titles_controller
...
  /movie/{id}:
    get:
      tags:
      - Titles
      summary: Get information about a Movie
      description: "Returns information about a particular Movie, including its title,\
        \ genre, director, actors, release date, and description."
      operationId: movie_id_get
      parameters:
      - name: id
        in: path
        description: ID of the Movie
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Information about the Movie/TV Show
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/movie'
        "400":
          description: Invalid request
        "404":
          description: Movie not found
      x-openapi-router-controller: swagger_server.controllers.titles_controller
  /tvshow/{id}:
    get:
      tags:
      - Titles
      summary: Get information about a TV Show
      description: "Returns information about a particular TV Show, including its\
        \ title, genre, director, actors, release date, and description."
      operationId: tvshow_id_get
      parameters:
      - name: id
        in: path
        description: ID of the TV Show
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Information about the Movie/TV Show
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tvshow'
        "400":
          description: Invalid request
        "404":
          description: TV Show not found
      x-openapi-router-controller: swagger_server.controllers.titles_controller
...
components:
  schemas:
    title:
      type: object
      properties:
        movie:
          $ref: '#/components/schemas/movie'
        tvshow:
          $ref: '#/components/schemas/tvshow'
      example:
        movie:
          duration: 2hrs 2mins
          cast:
          - Leonardo Di Caprio
          - Kate Winslet
          - Bradd Pitt
          plot: plot
          release_date: 2001-08-19T00:00:00.000+00:00
          director: Christopher Nolan
          genres:
          - Adventure
          - Comedy
          - Thriller
          rating: 9.36
          writer: Todd Phillips
          movie_id:
            id: 42
          title: Inception
          poster: https://fmd.com/poster.jpeg
        tvshow:
          duration: 55mins
          cast:
          - Leonardo Di Caprio
          - Kate Winslet
          - Bradd Pitt
          show_id:
            id: 42
          plot: plot
          release_date: 2001-08-19T00:00:00.000+00:00
          director: Christopher Nolan
          genres:
          - Adventure
          - Comedy
          - Thriller
          rating: 9.36
          season: 2
          writer: Todd Phillips
          title: Peaky Blinders
          poster: https://fmd.com/poster.jpeg
    movie_id:
      type: object
      properties:
        id:
          type: integer
          example: 42
      example:
        id: 42
    show_id:
      type: object
      properties:
        id:
          type: integer
          example: 42
      example:
        id: 42
...
    movie:
      type: object
      properties:
        movie_id:
          $ref: '#/components/schemas/movie_id'
        title:
          type: string
          example: Inception
        plot:
          maxLength: 500
          type: string
        cast:
          maxItems: 3
          type: array
          example:
          - Leonardo Di Caprio
          - Kate Winslet
          - Bradd Pitt
          items:
            type: string
        director:
          type: string
          example: Christopher Nolan
        writer:
          type: string
          example: Todd Phillips
        rating:
          type: number
          format: float
          example: 9.36
        genres:
          maxItems: 3
          minItems: 3
          type: array
          example:
          - Adventure
          - Comedy
          - Thriller
          items:
            type: string
        release_date:
          type: string
          format: date
          example: 2001-08-19
        duration:
          pattern: ^\d+hrs \d+mins$
          type: string
          example: 2hrs 2mins
        poster:
          type: string
          example: https://fmd.com/poster.jpeg
      example:
        duration: 2hrs 2mins
        cast:
        - Leonardo Di Caprio
        - Kate Winslet
        - Bradd Pitt
        plot: plot
        release_date: 2001-08-19T00:00:00.000+00:00
        director: Christopher Nolan
        genres:
        - Adventure
        - Comedy
        - Thriller
        rating: 9.36
        writer: Todd Phillips
        movie_id:
          id: 42
        title: Inception
        poster: https://fmd.com/poster.jpeg
    tvshow:
      type: object
      properties:
        show_id:
          $ref: '#/components/schemas/show_id'
        title:
          type: string
          example: Peaky Blinders
        plot:
          maxLength: 500
          type: string
        season:
          type: integer
          example: 2
        cast:
          maxItems: 3
          type: array
          example:
          - Leonardo Di Caprio
          - Kate Winslet
          - Bradd Pitt
          items:
            type: string
        director:
          type: string
          example: Christopher Nolan
        writer:
          type: string
          example: Todd Phillips
        rating:
          type: number
          format: float
          example: 9.36
        genres:
          maxItems: 3
          minItems: 3
          type: array
          example:
          - Adventure
          - Comedy
          - Thriller
          items:
            type: string
        release_date:
          type: string
          format: date
          example: 2001-08-19
        duration:
          pattern: ^\d+mins$
          type: string
          example: 55mins
        poster:
          type: string
          example: https://fmd.com/poster.jpeg
      example:
        duration: 55mins
        cast:
        - Leonardo Di Caprio
        - Kate Winslet
        - Bradd Pitt
        show_id:
          id: 42
        plot: plot
        release_date: 2001-08-19T00:00:00.000+00:00
        director: Christopher Nolan
        genres:
        - Adventure
        - Comedy
        - Thriller
        rating: 9.36
        season: 2
        writer: Todd Phillips
        title: Peaky Blinders
        poster: https://fmd.com/poster.jpeg
...